+++ src/search/score.rs
@@ -332,6 +332,7 @@ mod tests {
             description: None,
             deprecated: None,
             directives: vec![],
+            default: None,
             possible_types: vec![],
         }
     }