+++ src/semantic/cache.rs
@@ -618,6 +618,7 @@ mod tests {
             description: None,
             deprecated: None,
             directives: vec![],
+            default: None,
             possible_types: vec![],
             args: vec![],
         };